+++ inc/class-nettsmed-sentry.php
@@ -4,7 +4,7 @@ class NettsmedSentry {
 
 	private const DSN = 'https://c9255d6f2a9a6741ed0e264cc0d5fd3e@o4508484236607488.ingest.de.sentry.io/4510971168620624';
 	private const PLUGIN_DIR = 'nettsmed-support-plugin';
-	private const VERSION = '1.3.0';
+	private const VERSION = '1.9.0';
 
 	public static function init(): void {
 		if ( ! class_exists( '\Sentry\SentrySdk' ) ) {